bigint hardcoded mint

+++ frontend/js/app.js
@@ -286,7 +286,7 @@ function setTotalPrice() {
     return;
   }
   // const totalPriceWei = BigInt(info.deploymentConfig.mintPrice) * BigInt(mintInputValue);
-  const totalPriceWei = 10 * mintInputValue;
+  const totalPriceWei = BigInt(10 * mintInputValue);
   
   let priceType = '';
   if(chain === 'rinkeby' || chain === 'ethereum') {
